Adding a response bot
Adding a simple response bot that can handle single words and respond to them if the message entered is within a certain length (dont want to parse someones copy-pasta novel). Also having the response bot stop at the first message it sees, as I dont want the bot to spam with responses if multiple words are mentioned. And to exit